+++ resources/lang/de/validation.php
@@ -247,6 +247,9 @@ return [
         'icon_text'                         => 'Icon Text',
         'color'                             => 'Farbe',
         'allow_listing'                     => 'Raumsuche erlaubt',
-        'listed'                            => 'In Raumsuche einschließen'
+        'listed'                            => 'In Raumsuche einschließen',
+        'password_self_reset_enabled'       => 'Registrierten Nutzern ermöglichen das Passwort zurückzusetzen',
+        'timezone'                          => 'Zeitzone',
+        'default_timezone'                  => 'Standardzeitzone'
     ],
 ];