+++ resources/lang/en/validation.php
@@ -247,6 +247,9 @@ return [
         'icon_text'                         => 'Icon text',
         'color'                             => 'Color',
         'allow_listing'                     => 'Room search allowed',
-        'listed'                            => 'Include in room search'
+        'listed'                            => 'Include in room search',
+        'password_self_reset_enabled'       => 'Give registered users the possibility to reset their password',
+        'timezone'                          => 'Timezone',
+        'default_timezone'                  => 'Default timezone'
     ],
 ];