+++ tests/Feature/api/v1/SettingsTest.php
@@ -191,7 +191,9 @@ class SettingsTest extends TestCase
                 'link_target' => 'self',
                 'link_style'  => 'primary',
                 'icon'        => 'fas fa-door-open',
-            ]
+            ],
+            'password_self_reset_enabled' => '1',
+            'default_timezone'            => 'Europe/Berlin'
         ];
 
         $role       = factory(Role::class)->create();
@@ -217,7 +219,9 @@ class SettingsTest extends TestCase
                         'background' => '#4a5c66',
                         'link'       => 'http://localhost',
                         'icon'       => 'fas fa-door-open',
-                    ]
+                    ],
+                    'password_self_reset_enabled' => true,
+                    'default_timezone'            => 'Europe/Berlin'
                 ]
             ]);
     }
@@ -236,7 +240,9 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
-            'banner'                         => ['enabled' => false]
+            'banner'                         => ['enabled' => false],
+            'password_self_reset_enabled'    => false,
+            'default_timezone'               => 'Europe/Berlin'
         ];
 
         // Unauthorized Test
@@ -271,7 +277,9 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
-            'banner'                         => ['enabled' => false]
+            'banner'                         => ['enabled' => false],
+            'password_self_reset_enabled'    => '1',
+            'default_timezone'               => 'Europe/Berlin'
         ];
 
         // Unauthorized Test
@@ -309,7 +317,9 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
-            'banner'                         => ['enabled' => false]
+            'banner'                         => ['enabled' => false],
+            'password_self_reset_enabled'    => '1',
+            'default_timezone'               => 'Europe/Berlin'
         ];
 
         // Add necessary role and permission to user to update application settings
@@ -345,7 +355,9 @@ class SettingsTest extends TestCase
             'logo_file'                      => 'notimagefile',
             'pagination_page_size'           => 'notnumber',
             'own_rooms_pagination_page_size' => 'notnumber',
-            'room_limit'                     => 'notnumber'
+            'room_limit'                     => 'notnumber',
+            'password_self_reset_enabled'    => 'foo',
+            'default_timezone'               => 'timezone'
         ];
 
         $this->actingAs($this->user)->putJson(route('api.v1.application.update'), $payload)
@@ -360,7 +372,9 @@ class SettingsTest extends TestCase
                 'own_rooms_pagination_page_size',
                 'room_limit',
                 'banner',
-                'banner.enabled'
+                'banner.enabled',
+                'password_self_reset_enabled',
+                'default_timezone'
             ]);
 
         $payload = [
@@ -370,7 +384,9 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
-            'banner'                         => false
+            'banner'                         => false,
+            'password_self_reset_enabled'    => '1',
+            'default_timezone'               => 'Europe/Berlin'
         ];
 
         $this->putJson(route('api.v1.application.update'), $payload)